+++ lib/src/widgets/settings_panel/lit_settings_panel_controller.dart
@@ -17,15 +17,18 @@ class LitSettingsPanelController {
   /// dispose to ensure no [TickerProvider] leak will occur.  Call this method
   /// on the [StatefulWidget] by overriding its dispose method.
   void dispose() {
+    _listeners = [];
     animationController.dispose();
   }
 
   void dismissSettingsPanel() {
     animationController.reverse();
+    _callListerns();
   }
 
   void showSettingsPanel() {
     animationController.forward();
+    _callListerns();
   }
 
   bool get isShown {
@@ -38,5 +41,23 @@ class LitSettingsPanelController {
 
   bool get isAnimatingOrShown {
     return animationController.isAnimating || isShown;
+  }
+
+  /// A list currently added listeners.
+  List<Function> _listeners = [];
+
+  /// Calls all provided listeners one by one.
+  void _callListerns() {
+    for (Function func in _listeners) {
+      func();
+    }
+  }
+
+  /// Adds a listener to the panel.
+  ///
+  /// Each action ([dismissSettingsPanel],[showSettingsPanel]) will trigger the
+  /// provided callbacks.
+  void addListener(void Function() func) {
+    _listeners.add(func);
   }
 }